+++ src/context/SelectedTagContext.jsx
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+const SelectedTagContext = createContext({});
+
+export default SelectedTagContext;